+++ example/client/src/main/scala/samples/EitherSample.scala
@@ -8,15 +8,6 @@ import com.raquo.laminar.api.L.*
 
 import com.raquo.airstream.state.Var
 
-case class Cat(name: String, weight: Int)
-case class Dog(name: String, weight: Int)
-
-given Defaultable[Cat] with
-  def default = Cat("", 0)
-
-given Defaultable[Dog] with
-  def default = Dog("", 0)
-
 case class EitherSample(
     either: Either[Cat, Dog],
     optionalInt: Option[Int]
