+++ example/client/src/main/scala/samples/package.scala
@@ -1,0 +1,12 @@
+package samples
+
+import dev.cheleb.scalamigen.Defaultable
+
+case class Cat(name: String, weight: Int)
+case class Dog(name: String, weight: Int)
+
+given Defaultable[Cat] with
+  def default = Cat("", 0)
+
+given Defaultable[Dog] with
+  def default = Dog("", 0)
